requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.2
click>=8.1.3
python-dotenv>=1.0.0
requests>=2.30.0
psycopg2-binary>=2.9.5
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
matplotlib>=3.7.0
reportlab>=3.6.0
Flask-Mail>=0.9.1 